Fixed bugs in the gc interval calculations and explained them better. Afaics those bugs could not have any severe effects, though; the only one I've seen is that there would be an average of 16.67% garbage in each gc instead of 20%.

+++ src/gc.h
@@ -254,7 +254,7 @@ int gc_cycle_push(void *x, struct marker *m, int weak);
 void do_gc_recurse_svalues(struct svalue *s, int num);
 void do_gc_recurse_short_svalue(union anything *u, int type);
 int gc_do_free(void *a);
-int do_gc(void);
+size_t do_gc(void);
 void f__gc_status(INT32 args);
 void cleanup_gc(void);
 /* Prototypes end here */
